+++ include/common/monolish_crs.hpp
@@ -47,6 +47,20 @@ private:
    */
   mutable bool gpu_status = false;
 
+  /**
+   * @brief hash, created from row_ptr and col_ind
+   */
+  size_t structure_hash;
+
+  /**
+   * @brief create index array hash (to compare structure)
+   * @note
+   * - # of computation: nnz + rowN + 1
+   * - Multi-threading: true
+   * - GPU acceleration: true
+   */
+  void create_hash();
+
 public:
   /**
    * @brief CRS format value, which stores values of the non-zero elements (size
@@ -202,6 +216,13 @@ public:
    * - GPU acceleration: false
    **/
   std::string type() const { return "CRS"; }
+
+  /**
+   * @brief get index array hash (to compare structure)
+   * @note
+   * - # of computation: 1
+   */
+  size_t get_hash() { return structure_hash; }
 
   // communication
   // ///////////////////////////////////////////////////////////////////////////